Move bit_set to new file
