+++ rpcs3/Emu/Cell/Modules/cellSsl.cpp
@@ -7,6 +7,7 @@
 #include "Utilities/File.h"
 #include "Emu/VFS.h"
 #include "Emu/IdManager.h"
+#include "util/bit_set.hpp"
 
 #include "cellRtc.h"
